reorganize search input
